Set BUP_MAIN_EXE and BUP_DIR in testpacks()
testpacks did not set the BUP_MAIN_EXE and BUP_DIR environment variables
before invoking git.PackWriter(), which caused the wvtest to be run
looking for the wrong files.

This is currently Cygwin-specific problem, but the all other tests in this
series set these environment variables before running any other logic,
but for a currently unknown reason the environment does not appear to be
preserved properly under cygwin. With this change, this test executes
successfully while still being a correct test.

Signed-off-by: Will Rouesnel <w.rouesnel@gmail.com>