Add -e,--remote-shell support to remote commands.
Added support for the rsync-style -e command line option to explicitly
specify the command line bup should use for the remote shell. This
allows the use of advanced SSH options by specifying them on the command
line ala:

	bup init -r user@remotehost -e 'ssh -oCompression=yes'

or the use of alternate shells which accept addressing and command
invocation in the SSH style.

Signed-off-by: Will Rouesnel <w.rouesnel@gmail.com>