Procedurally generate t/sampledata
  * Remove existing sampledata from t/
  * Initial check-in of generate-test-data.sh script.
  * Add make testdata dependency to the make runtests target.
  * Remove testdata when running make clean.
  * Add t/sampledata and t/testfile1, t/testfile2 to .gitignore.

  * Fixes symlink tests in tindex.py not working on Cygwin.

Storing the sample data for tests in the Git repository causes problems
with cross-platform use due to the differing handling of symlinks by
Git, and the very different behavior that can arise when using symlinks
to files and symlinks to directories. This is particularly a problem on
Cygwin.

This commit resolves this problem by replacing the fixed sample data
with a shell script (generate-test-data.sh) which dynamically creates
suitable test data before tests are run.

Signed-off-by: Will Rouesnel <w.rouesnel@gmail.com>