Add atomic rename support to bup-index
This commit adds preliminary support to bup-index for atomically
renaming files in the index.

This allows an inotify or similar application to detect atomic rename
operations, and request the index be updated to reflect the new file
location without a full rehash of the file.

Files copied moved into the directory tree are detected and added by the
normal update_index() function.