Add restore-files command to bup-server for tar file piping
This patch adds the restore-files command to bup-server. This enables
bup-server to dump the contents of a given path within its working bup
repository as a tar-formatted (and gzip/bz2 compressed) archive written
to stdout.

This command works *without* any additional protocol communication, and
can be activated directly over SSH - i.e.
	echo "restore-files --transfermode=tar /backup/latest" | \
	ssh user@host bup-server > myarchive.tar

Please note it *depends* on the quiet-mode patch for this functionality.

tserver.py tests the functionality of the bup-server.

Since the bup-server is relatively complex, we need a Python script in
order to access the bup-archive and filesystem directly to perform
byte-by-byte tests and comparisons.