Add restore_files_iter() function to vfs.py
Provides support for creating a generator which will iterate over a
given backup set and return all nodes in a top-down order.

The function is top-down with respect to directory structure, but files for
each level are returned first.

The function accepts a patchstack parameter (intended for the
protocol.Pathstack structure) in order to track the differential path name of
each returned node. If "None" is provided, this element is ignored.