Add File.hashlist() and local_shalist() functions to vfs.py Add tvfs.py to provide test coverage for VFS functions.
This patch adds a function and helpers to return the hashlist of a given
VFS node without returning or even accessing the binary blob data
itself.

It also adds the local_shalist() function, which applies the bup
hashsplit algorithm to a give *real* file path but returns only the
hashlist which would have been generated as (offset, hash, size) tuples.

note: _sha_iter() does not return the size of the binary blob, as this
would require reading the binary blob in order to determine it.

Both these functions are necessary to support differential file
transfers from bup-server.

vfs.py has very little specific test coverage. Since the new hashsplit
functions are concerned with data integrity, this patch adds a test
module for the vfs to ensure we can do function level testing.