Add 'bup' differential transfermode support.
This patch adds the elements needed for bup-server to support providing
differential or dumb interactive file transfers to a client.

The 'bup' transfermode is added as a supported RestoreSessionRequest
mode.

Clients can request data by sending metadata requests, and request
hashlists and blobs by sending efficient PipelineRequests which ask for
the data in the sequence the server remembers sending it.

This can be used to skip-over iteration and to do differential transfers
of files and data the client already has.

Add testcases for restore-files --transfermode=bup to tserver.py

This patch also adds initial testcases for the 'bup' transfermode of
restore-files to tserver.py.

The initial test case initializes a dummy bup repository of testdata,
and starts a server to request data from then does byte-by-byte
comparison of received data to the actual files on the disk.