Don't delete previously updated paths in the same index run.
This patch forces update_index to keep track of the root paths it has
previously written to. If a subsequent, root somehow excludes a previous
one, update_index is prevented from marking the previously updated files
as deleted.

By itself this change has no effect, but it is necessary for the
improved xdev shadowing and subsequent patches to work.