Do not reduce paths on separate devices when -x is enabled
This patch stops bup index -x/--xdev/--one-file-system from reducing
paths in the same hierarchy if they are on separate devices.

Previously bup index's reduce_paths function would look at a /usr and
/usr/local and remove /usr/local since it is included by /usr.

However if /usr and /usr/local are on different devices and -x is set,
then it would remove /usr/local but would fail to index it.

This patch causes the reduce_paths() function involved to check that a
path marked to be reduced is on the *same* device *if* -x is enabled.

This solves the problem and produces expected functionality.

This patch includes a relevant test in test-xdev.sh to confirm
functionality.