reduce_paths() check directories between folded paths are traversable
This patch causes reduce_paths() to check that all directories between a
candidate to be reduced and the path it will be reduced to can have
their contents listed.

This ensures drecurse is able to descend into these directories then
read permissions are not available to the current user/group that bup is
running as.

If a directory cannot be traversed, then it is not reduced and is backed
up explicitly.

For example:
	/home/user/root/www

Supposing /home/user/root does not have read permissions
(i.e. 0111 permissions) but the user specifies:

	bup index -u /home/user /home/user/root/www

Then bup will correctly retain the /home/user/root/www path that it
can access, and index it appropriately.