Dynamic VFS view of bup-repositories.
This patch adds support for efficiently updating the VFS's view of a bup
repository, enabling client applications like bup-fuse and bup-web to
show new commits and branches as they added to the backend repository.

It accomplishes this by checking the the repository for new
branches/commits when a root node is resolved, and patching the view
rather then discarding it. This retains cached metadata in memory and
avoids slowdowns due to file-size calculations.

Any VFS-using bup command will automatically take advantage of this
feature. This principally enables exporting a bup-fuse or bup-web view
as a persistent service.