Modify metadata to support storing the git file mode.
In the advent that a VFS node has no metadata, it would be nice not to
break the normal metadata features. This patch adds a new tag _rec_vfs
which stores the git file mode which the VFS falls back on.

This tag will normally never be used unless explicitly needed, so
there is minimal overhead (since files will normally have much larger
metadata anyway).